@@ -1,12 +1,14 @@
-#!/usr/bin/env node
-import { readdirSync, lstatSync } from 'fs';
+import chalk from 'chalk';
+import chokidar from 'chokidar';
+import minimist from 'minimist';
 import path from 'path';
 import { createFilter } from '@rollup/pluginutils';
+import { readdirSync, lstatSync, readFileSync } from 'fs';
 import buildspec from '../lib/buildspec';
 import { bundler, openPage } from '../lib/bundler';
 import { getBrowser } from '../lib/puppeteer';
-import { cyan, yellow, green, red } from 'chalk';
-import chokidar from 'chokidar';
+
+const { cyan, yellow, green, red } = chalk;
 
 const cmd = process.argv[2];
 
@@ -24,7 +26,14 @@ let rollupWatcher = null;
 
 const targetPath = process.argv[3] || '.';
 
-console.log('');
+const cmdArgs = minimist(process.argv.slice(3), {
+	boolean: ['build'],
+	alias: {
+		b: 'build',
+		u: 'url',
+		n: 'name',
+	},
+});
 
 // More graceful exit
 process.on('SIGINT', () => {
@@ -53,7 +62,7 @@ switch (cmd) {
  * @param {string} targetPath
  */
 async function buildSingleEntry(targetPath) {
-	const [config] = getMatchingBuildspec(targetPath);
+	const [config] = await getMatchingBuildspec(targetPath);
 
 	if (!config) {
 		console.log(red("Couldn't find buildspec.json for the variant", targetPath));
@@ -107,7 +116,7 @@ async function buildSingleEntry(targetPath) {
  */
 async function buildMultiEntry(targetPath) {
 	const buildOnly = cmd === 'build-all';
-	const buildspecs = getMatchingBuildspec(targetPath);
+	const buildspecs = await getMatchingBuildspec(targetPath);
 	const [testConfig] = buildspecs;
 
 	if (!testConfig) {
@@ -160,7 +169,14 @@ async function buildMultiEntry(targetPath) {
  * @param {string} targetPath
  */
 async function createScreenshots(targetPath) {
-	const buildspecs = getMatchingBuildspec(targetPath);
+	const buildspecs = await getMatchingBuildspec(targetPath);
+
+	const { url: cmdArgUrl, name, ...screenshotArgs } = cmdArgs;
+	let cmdArgBuild = false;
+	if ('build' in screenshotArgs) {
+		cmdArgBuild = true;
+		delete screenshotArgs.build;
+	}
 
 	if (!buildspecs.length) {
 		console.log(red('0 test variants found with path:'));
@@ -170,37 +186,134 @@ async function createScreenshots(targetPath) {
 
 	let origPage;
 
+	console.log(cyan(`Taking screenshots`));
+	console.log();
+
 	for (const config of buildspecs) {
 		const nth = buildspecs.indexOf(config) + 1;
-		const { entryFile, entryFileExt } = config;
+		if (cmdArgUrl) {
+			if (isNaN(cmdArgUrl)) {
+				config.url = [cmdArgUrl];
+			} else {
+				if (Array.isArray(config.url)) {
+					const urlIndex = +cmdArgUrl;
+					if (urlIndex < config.url.length) {
+						config.url = [config.url[urlIndex]];
+					} else {
+						console.log(yellow(`Undefined index for test config url. Argument was`), '--url=' + cmdArgUrl);
+						console.log(yellow(`Current config`), config.url);
+					}
+				} else {
+					console.log(
+						yellow(`Test config url wasn't an array, can't use indexed url argument. Argument was`),
+						'--url=' + cmdArgUrl
+					);
+				}
+			}
+		}
+		const { testPath, buildDir, entryFile, entryFileExt, screenshot = {}, onLoad, onBefore } = config;
+
+		Object.assign(screenshot, screenshotArgs);
+
+		let {
+			waitFor,
+			waitForOptions = {},
+			onLoad: screenshotOnLoad,
+			onBefore: screenshotOnBefore,
+			...pptrScreenshotOptions
+		} = screenshot;
 		const entryName = path.basename(entryFile, '.' + entryFileExt);
-		const output = await bundler({ ...config, preview: true });
+
+		// Bundle main events and screenshot events
+		const singleOnLoad = async (page) => {
+			if (onLoad) await onLoad(page);
+			if (screenshotOnLoad) await screenshotOnLoad(page);
+		};
+
+		const singleOBefore = async (page) => {
+			if (onBefore) await onBefore(page);
+			if (screenshotOnBefore) await screenshotOnBefore(page);
+		};
+
+		const bundleConfig = { ...config, onLoad: singleOnLoad, onBefore: singleOBefore, preview: true };
+
+		if (!cmdArgBuild) {
+			const bundlePath = path.join(testPath, buildDir, `${entryName}.bundle.js`);
+			if (lstatSync(bundlePath).isFile()) {
+				bundleConfig.assetBundle = {
+					bundle: readFileSync(bundlePath).toString(),
+				};
+			}
+		}
+
+		const output = bundleConfig.assetBundle ? bundleConfig : await bundler(bundleConfig);
+
+		console.log(cyan(`Creating a bundle for screenshot`), entryFile);
+		console.log();
+
 		const page = await openPage({ ...output, headless: true, devtools: false });
 		const url = Array.isArray(config.url) ? config.url[0] : config.url;
+
+		const waitForAll = async (page) => {
+			if (typeof waitFor === 'string') {
+				console.log(cyan(`Waiting for selector ${waitFor}...`));
+				await page.waitForSelector(waitFor, waitForOptions);
+			}
+			if (typeof waitFor === 'number') {
+				console.log(cyan(`Waiting for timeout ${waitFor} ms...`));
+				await page.waitForTimeout(waitFor, waitForOptions);
+			}
+			if (typeof waitFor === 'function') {
+				console.log(cyan(`Waiting for function...`));
+				await page.waitForFunction(waitFor, waitForOptions);
+			}
+		};
+
+		if (pptrScreenshotOptions.fullPage === undefined) {
+			pptrScreenshotOptions.fullPage = true;
+		}
+		if (pptrScreenshotOptions.clip) {
+			pptrScreenshotOptions.fullPage = false;
+		}
+
+		await waitForAll(page);
+
 		// Take screenshot from variant
 		await page.screenshot({
-			path: path.join(config.testPath, config.buildDir, `screenshot-${entryName}-v${nth}.png`),
-			fullPage: true,
+			...pptrScreenshotOptions,
+			path: path.join(config.testPath, config.buildDir, `screenshot-${path.basename(name || entryName)}-v${nth}.png`),
 		});
+
+		console.log(cyan(`Screenshot ready`), `${entryFile}, variant ${nth}`);
+
 		// Get new page for the original (without listeners etc)
-		if (!origPage) {
-			origPage = await page.browser().newPage();
-		}
+		origPage = await page.browser().newPage();
+
+		await singleOBefore(origPage);
+
 		// Go to the same url and take the screenshot from the original as well.
 		await origPage.goto(url, { waitUntil: 'networkidle0' });
+
+		await singleOnLoad(origPage);
+
+		await waitForAll(origPage);
+
 		await origPage.screenshot({
-			path: path.join(config.testPath, config.buildDir, `screenshot-${entryName}-orig.png`),
-			fullPage: true,
+			...pptrScreenshotOptions,
+			path: path.join(config.testPath, config.buildDir, `screenshot-${path.basename(name || entryName)}-orig.png`),
 		});
+
+		console.log(cyan(`Screenshot ready`), `${entryFile}, original`);
+		console.log();
+
 		console.log(green('Took screenshots for'), entryFile.replace(process.env.INIT_CWD, ''));
 		console.log();
 	}
 
+	console.log(green('Done.'));
 	if (origPage) {
 		await origPage.browser().close();
 	}
-
-	console.log(green('Done.'));
 }
 
 /**
@@ -208,7 +321,7 @@ async function createScreenshots(targetPath) {
  * @param {string} targetPath
  * @return {Object[]}
  */
-function getMatchingBuildspec(targetPath) {
+async function getMatchingBuildspec(targetPath) {
 	let indexFiles = [];
 
 	targetPath = path.resolve(process.env.INIT_CWD, targetPath);
@@ -225,8 +338,8 @@ function getMatchingBuildspec(targetPath) {
 		);
 	}
 
-	return indexFiles
-		.map((entryFile) => {
+	const entries = await Promise.all(
+		indexFiles.map(async (entryFile) => {
 			const filter = createFilter([/\.(jsx?|tsx?|(le|sa|sc|c)ss)$/]);
 			if (entryFile.includes('.build')) {
 				return null;
@@ -234,11 +347,13 @@ function getMatchingBuildspec(targetPath) {
 			if (!filter(entryFile) && !lstatSync(entryFile).isDirectory()) {
 				return null;
 			}
-			const spec = buildspec(entryFile);
+			const spec = await buildspec(entryFile);
 			if (spec && new RegExp(`${spec.buildDir}(/|$)`).test(entryFile)) {
 				return null;
 			}
 			return spec;
 		})
-		.filter(Boolean);
+	);
+
+	return entries.filter(Boolean);
 }
