@@ -2,6 +2,11 @@ import path from 'path';
 import fs from 'fs';
 import { createFilter } from '@rollup/pluginutils';
 import merge from 'lodash.merge';
+import { createRequire } from 'module';
+import { convertToEsmPath } from './utils';
+
+// Require is not defined in ES module scope
+const specRequire = createRequire(import.meta.url);
 
 /**
  * Config defaults
@@ -18,24 +23,35 @@ export const config = {
 	sourcemap: true,
 	watch: false,
 	windowSize: [1920, 1080],
+	features: {
+		hooks: 'auto',
+		jsx: 'auto',
+		classes: true,
+		className: true,
+		namespaces: true,
+	},
 };
 
-function getConfigFileJsonOrJSContent(configPath) {
+async function getConfigFileJsonOrJSContent(configPath) {
 	const fileDir = path.dirname(configPath);
 	const fileWithoutExt = path.basename(configPath, '.json');
 	const configPathJs = path.resolve(fileDir, fileWithoutExt + '.js');
 	if (fs.existsSync(configPath)) {
-		delete require.cache[configPath];
-		return require(configPath);
+		// Clear require cache before loading the file
+		delete specRequire.cache[configPath];
+		return specRequire(configPath);
 	} else if (fs.existsSync(configPathJs)) {
-		delete require.cache[configPathJs];
-		return require(configPathJs);
+		// Import file with a cache buster
+		const { default: configMod } = await import(
+			convertToEsmPath(configPathJs) + '?cb=' + Math.random().toString(36).substring(3)
+		);
+		return configMod;
 	}
 	return {};
 }
 
-function mergeSpecs(testConfig, specConfigPath) {
-	const specConfig = getConfigFileJsonOrJSContent(specConfigPath);
+async function mergeSpecs(testConfig, specConfigPath) {
+	const specConfig = await getConfigFileJsonOrJSContent(specConfigPath);
 	if (!Object.keys(specConfig).length) {
 		return;
 	}
@@ -44,7 +60,7 @@ function mergeSpecs(testConfig, specConfigPath) {
 	testConfig._specFiles.push(specConfigPath, specConfigPath.substr(0, specConfigPath.length - 2));
 }
 
-export default function (testPath) {
+export default async function (testPath) {
 	if (!testPath) {
 		console.log('Test folder is missing');
 		process.exit();
@@ -72,11 +88,11 @@ export default function (testPath) {
 	// Check & load build related data
 	const testConfig = {};
 
-	mergeSpecs(testConfig, path.resolve(process.cwd(), 'config.json'));
+	await mergeSpecs(testConfig, path.resolve(process.cwd(), 'config.json'));
 
 	try {
-		mergeSpecs(testConfig, path.join(testPath, '..', 'buildspec.json'));
-		mergeSpecs(testConfig, path.join(testPath, 'buildspec.json'));
+		await mergeSpecs(testConfig, path.join(testPath, '..', 'buildspec.json'));
+		await mergeSpecs(testConfig, path.join(testPath, 'buildspec.json'));
 	} catch (error) {
 		console.error(error);
 	}
@@ -120,5 +136,13 @@ export default function (testPath) {
 		testConfig.url = [testConfig.url];
 	}
 
-	return { ...config, ...testConfig, testPath, entryFile, entryPart, entryFileExt: entryFile.split('.').pop() };
+	return {
+		...config,
+		...testConfig,
+		features: { ...config.features, ...(testConfig.features || {}) },
+		testPath,
+		entryFile,
+		entryPart,
+		entryFileExt: entryFile.split('.').pop(),
+	};
 }
