@@ -1,38 +1,41 @@
-import path from 'path';
-import fs from 'fs';
-import { rollup } from 'rollup';
-import { watch as rollupWatch } from 'rollup';
-import nodeResolve from '@rollup/plugin-node-resolve';
-import babelPlugin from '@rollup/plugin-babel';
-import { terser } from 'rollup-plugin-terser';
-import styles from 'rollup-plugin-styles';
-import stringHash from 'string-hash';
-import autoprefixer from 'autoprefixer';
-import alias from '@rollup/plugin-alias';
-import replace from '@rollup/plugin-replace';
-import commonjs from '@rollup/plugin-commonjs';
-import { openPage } from './puppeteer';
-import inlineSvg from 'rollup-plugin-inline-svg';
-import svgImport from 'rollup-plugin-svg-hyperscript';
-import image from '@rollup/plugin-image';
-import chunkImage from './plugins/rollup-chunk-image';
-import browserslist from 'browserslist';
-import { getFlagEnv, hashf, unifyPath } from './utils';
-import chalk from 'chalk';
-import rimraf from 'rimraf';
-import { glob } from 'glob';
-import { preactDebug } from './preact-debug-plugin';
-import typescript from 'rollup-plugin-typescript2';
-import iife from 'rollup-plugin-iife';
-import { minify as terserMinify } from 'terser';
-
 export * from './puppeteer';
+import alias from '@rollup/plugin-alias';
+import autoprefixer from 'autoprefixer';
+import chalk from 'chalk';
+import commonjs from '@rollup/plugin-commonjs';
+import fs from 'fs';
+import glob from 'glob';
+import iife from 'rollup-plugin-iife';
+import image from '@rollup/plugin-image';
+import inlineSvg from 'rollup-plugin-inline-svg';
+import nodeResolve from '@rollup/plugin-node-resolve';
+import path from 'path';
+import replace from '@rollup/plugin-replace';
+import rimraf from 'rimraf';
+import stringHash from 'string-hash';
+import styles from 'rollup-plugin-styles';
+import svgImport from 'rollup-plugin-svg-hyperscript';
+import typescript from 'rollup-plugin-typescript2';
+import { babel } from '@rollup/plugin-babel';
+import { fileURLToPath } from 'url';
+import { minify as terserMinify } from 'terser';
+import { rollup } from 'rollup';
+import { terser } from 'rollup-plugin-terser';
+import { watch as rollupWatch } from 'rollup';
+import browserslist from 'browserslist';
+import chunkImage from './plugins/rollup-chunk-image';
+import { getFlagEnv, hashf, unifyPath } from './utils';
+import { openPage } from './puppeteer';
+import { preactDebug } from './preact-debug-plugin';
+
+// __dirname is not defined in ES module scope
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
 
 const browsers = browserslist();
 
 const supportIE = !!browsers.find((b) => b.startsWith('ie'));
 
-const rootDir = path.join(__dirname, '..', '..');
+const rootDir = __dirname.replace(/(\/|\\)(dist|bin|lib).*/, '');
 
 export async function bundler(testConfig) {
 	let { entryFile, testPath, entry, entryPart, preview = false } = testConfig;
@@ -70,6 +73,7 @@ export async function bundler(testConfig) {
 		chunks,
 		chunkImages,
 		watch,
+		features,
 		bundler: bundlerConfig = {},
 	} = testConfig;
 	const minify = configMinify ?? !watch;
@@ -140,6 +144,14 @@ export async function bundler(testConfig) {
 	process.env.TEST_ENV = TEST_ENV;
 	process.env.TEST_ID = TEST_ID;
 
+	const featuresReplaces = {};
+	Object.entries(features).forEach(([key, value]) => {
+		if (value !== 'auto') {
+			const letter = key === 'className' ? 'x' : key[0];
+			featuresReplaces[`config.${letter}`] = value.toString();
+		}
+	});
+
 	const chunksInputConfig = chunks
 		? {
 				preserveEntrySignatures: 'allow-extension',
@@ -149,7 +161,13 @@ export async function bundler(testConfig) {
 
 	const inputOptions = {
 		input: [entryFile],
-		treeshake: 'smallest',
+		treeshake: {
+			propertyReadSideEffects: false,
+			moduleSideEffects: true,
+			tryCatchDeoptimization: false,
+			unknownGlobalSideEffects: false,
+			correctVarValueBeforeDeclaration: false,
+		},
 		plugins: getPluginsConfig(
 			[
 				['preact-debug'],
@@ -224,6 +242,7 @@ export async function bundler(testConfig) {
 							'process.env.PREVIEW': PREVIEW,
 							'process.env.TEST_ENV': TEST_ENV,
 							'process.env.TEST_ID': JSON.stringify(TEST_ID),
+							...featuresReplaces,
 						},
 					},
 				],
@@ -386,15 +405,18 @@ export async function bundler(testConfig) {
 			sections: [],
 		};
 
-		// Create javascript bundle
+		// Create javascript bundle and keep track where chunk code starts for sourcemaps
 		let assetBundle = outputChunks
 			.filter((c) => !!c.code)
 			.reduce((bundle, chunk) => {
 				if (chunk.map) {
 					bundleMap.sections.push({ offset: { line: bundle.split('\n').length - 1, column: 0 }, map: chunk.map });
 				}
-				return `${bundle.trim()}\n${chunk.code}`;
-			}, '');
+				let bundleCode = bundle.trim();
+				if (bundleCode) bundleCode += '\n';
+				return `${bundleCode}${chunk.code}`;
+			}, '')
+			.trim();
 
 		if (styles) {
 			/* prettier-ignore */
@@ -539,7 +561,11 @@ export async function bundler(testConfig) {
 				}
 				bundleOutput = await event.result.generate(outputOptions);
 				assetBundle = await createAssetBundle(buildDir, entryPart);
-				page = await openPage({ ...testConfig, assetBundle }, true);
+				try {
+					page = await openPage({ ...testConfig, assetBundle }, true);
+				} catch (error) {
+					console.log('Error while opening page', error);
+				}
 				hasError = false;
 				await event.result.close();
 			}
@@ -577,7 +603,7 @@ function getPluginsConfig(defaults, override = []) {
 	const fns = {
 		alias,
 		'node-resolve': nodeResolve,
-		babel: babelPlugin,
+		babel,
 		styles,
 		replace,
 		commonjs,
