@@ -1,6 +1,8 @@
-import { yellow, green } from 'chalk';
+import chalk from 'chalk';
 import puppeteerCore from 'puppeteer-core';
 import { getFlagEnv } from './utils';
+
+const { yellow, green } = chalk;
 
 const isTest = getFlagEnv('TEST_ENV');
 let initial = true;
@@ -15,11 +17,21 @@ let aboutpage = null;
  * @param {Boolean} [singlePage]
  */
 export async function openPage(config, singlePage) {
-	const { url: urls, assetBundle } = config;
-	const url = getDefaultUrl(urls);
+	const { url: urls, assetBundle, onBefore, onLoad } = config;
+	let url = getDefaultUrl(urls);
 	let urlAfterLoad = url;
 
 	const page = await getPage(config, singlePage);
+	if (isOneOfBuildspecUrls(page.url(), urls)) {
+		url = page.url();
+	}
+
+	// Remove previous listeners
+	page.removeAllListeners();
+
+	if (onBefore) {
+		await onBefore(page);
+	}
 
 	const pageTags = [];
 
@@ -36,22 +48,20 @@ export async function openPage(config, singlePage) {
 		console.log(
 			...[`#${++counter}`, text, `page ${yellow(url)}`].concat(pageTags.length ? ['with custom:', pageTags] : [])
 		);
-	} else {
-		page.on('console', (message) => console.log(message.text())).on('pageerror', ({ message }) => console.log(message));
+	}
+	if (isTest || config.debug) {
+		page
+			.on('console', (message) => console.log('LOG: ', message.text()))
+			.on('pageerror', ({ message }) => console.log('ERR: ', message));
 	}
 
 	if (aboutpage) {
-		aboutpage.close();
+		await aboutpage.close();
 		aboutpage = null;
 	}
 
 	// TODO: Change to something less error prone if possible. Fixes the "already exposed function" error
 	page._pageBindings.set('isOneOfBuildspecUrls', (url) => isOneOfBuildspecUrls(url, urls));
-
-	// Remove previous listener
-	if (loadListener) {
-		page.off('domcontentloaded', loadListener);
-	}
 
 	// Add listener for load event. Using event makes it possible to refresh the page and keep these updates.
 	loadListener = async () => {
@@ -155,6 +165,10 @@ export async function openPage(config, singlePage) {
 	page.on('domcontentloaded', loadListener);
 
 	await page.goto(url, { waitUntil: 'networkidle0' });
+
+	if (onLoad) {
+		await onLoad(page);
+	}
 
 	initial = false;
 	// There might be some redirect/hash/etc
@@ -289,6 +303,8 @@ export async function getPage(config, singlePage) {
 		console.log('\n\x1b[31m%s\x1b[0m', 'Page error!');
 		console.error(pageerr);
 	});
+
+	await newPage.setDefaultNavigationTimeout(0);
 
 	return newPage;
 }
