@@ -1,6 +1,8 @@
-import { append, pollQuerySelector } from 'a-b-doer';
+import { append, pollQuerySelector, unmount } from 'a-b-doer';
 import { Simple, RefHook, Hooks, Switch, OrderApp } from './templates';
 import './styles.scss';
+
+window.unmount = unmount;
 
 pollQuerySelector('html body', (body) => {
 	append(<Simple id="tpl1" foo="0" />, body);
@@ -8,7 +10,7 @@ pollQuerySelector('html body', (body) => {
 	append(<Simple id="tpl3" foo="2" bar="1" />, body);
 	append(<RefHook id="tpl4" />, body);
 	append(<Hooks id="tpl5" />, body);
-	append(<Hooks id="tpl6" />, body);
+	window.hookVnode = append(<Hooks id="tpl6" />, body);
 	append(<Switch id="tpl7" />, body);
 	const Title = document.querySelector('h1');
 	append(<Title id="tpl8" data-o="h1" class="title" />, body);
